Added question complexity metric
Added a metric that measures the complexity of just the question asked after parsing.
It uses the same regex used to measure the complexity of a string in complexity.py
I added this mostly because I think it's interesting to have an additional complexity metric that doesn't care how the python code for the problem is written, just what questions get asked.